Root driver rename
