+++ code/system/sys.fs.indexeddb/src/IndexedDb.Driver/index.mts
@@ -1,1 +1,1 @@
-export * from './FsDriver.mjs';
+export * from './FsIndexedDbDriver.mjs';
